+++ resources/helpers/EmailHelper.php
@@ -2,9 +2,10 @@
 namespace App\Helpers;
 
 use App\Helpers\ViewHelper;
-use App\Mail\ForgottenPassword;
 use App\Mail\NoteNotification;
+use App\Mail\ForgottenPassword;
 use Illuminate\Support\Facades\Mail;
+use App\Mail\TransactionNotification;
 use Illuminate\Mail\Mailables\Address;
 
 class EmailHelper
@@ -20,6 +21,11 @@ class EmailHelper
             case 'note.notification':
                 Mail::to($to)->send(
                     new NoteNotification($snippets)
+                );
+                break;
+            case 'transaction.notification':
+                Mail::to($to)->send(
+                    new TransactionNotification($snippets)
                 );
                 break;
             default: